% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agroclimatico-package.R
\docType{package}
\name{agroclimatico-package}
\alias{agroclimatico}
\alias{agroclimatico-package}
\title{agroclimatico: Índices y Estadísticos Climáticos e Hidrológicos}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Conjunto de funciones para calcular índices y estadísticos climáticos hidrológicos a partir de datos tidy. Incluye una función para graficar resultados georeferenciados y e información cartográfica.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/AgRoMeteorologiaINTA/agroclimatico}
  \item \url{https://agrometeorologiainta.github.io/agroclimatico/}
  \item Report bugs at \url{https://github.com/AgRoMeteorologiaINTA/agroclimatico/issues}
}

}
\author{
\strong{Maintainer}: Paola Corrales \email{paola.corrales@cima.fcen.uba.ar}

Authors:
\itemize{
  \item Elio Campitelli \email{elio.campitelli@cima.fcen.uba.ar} (\href{https://orcid.org/0000-0002-7742-9230}{ORCID})
  \item Natalia Gattinoni \email{gattinoni.natalia@inta.gob.ar}
}

Other contributors:
\itemize{
  \item Yanina Bellini Saibene \email{yabellini@gmail.com} [contributor]
  \item  INTA [copyright holder]
  \item Ruida Zhong [copyright holder]
}

}
\keyword{internal}
