% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agromet-package.R
\docType{package}
\name{agromet-package}
\alias{agromet}
\alias{agromet-package}
\title{agromet: Índices y Estadísticos Climáticos e Hidrológicos}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Conjunto de funciones para calcular índices y estadísticos climáticos 
    hidrológicos a partir de datos tidy. Incluye una función para graficar resultados 
    georeferenciados y e información cartográfica.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/AgRoMeteorologiaINTA/agromet}
  \item Report bugs at \url{https://github.com/AgRoMeteorologiaINTA/agromet/issues}
}

}
\author{
\strong{Maintainer}: Yanina Bellini Saibene \email{yabellini@gmail.com} [contributor]

Authors:
\itemize{
  \item Elio Campitelli \email{elio.campitelli@cima.fcen.uba.ar} (\href{https://orcid.org/0000-0002-7742-9230}{ORCID})
  \item Paola Corrales \email{paola.corrales@cima.fcen.uba.ar}
  \item Natalia Gattinoni \email{gattinoni.natalia@inta.gob.ar}
}

Other contributors:
\itemize{
  \item  INTA [copyright holder]
}

}
\keyword{internal}
