% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/informes.R
\name{agromet_informe}
\alias{agromet_informe}
\title{Formatos de salida para Informes}
\usage{
agromet_informe(..., toc = TRUE)
}
\arguments{
\item{...}{cualquier argumento que requiera \code{\link[rmarkdown:pdf_document]{rmarkdown::pdf_document()}}.}
}
\description{
Los formatos utilizan como base \code{\link[rmarkdown:pdf_document]{rmarkdown::pdf_document()}} y una plantilla
específica de LaTex.
}
