% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spi.R
\name{completar_serie}
\alias{completar_serie}
\title{Completa una serie de datos}
\usage{
completar_serie(datos, fecha, resolucion)
}
\arguments{
\item{datos}{Tabla (data.frame, data.table, tibble) a completar.}

\item{fecha}{Columna de fecha.}

\item{resolucion}{Texto para definir la resolución puede ser
cualquier valor acpetado por el argumento \code{by} de la función \code{\link[=seq.Date]{seq.Date()}}
o su traducción al español. Es decir, un valor que contanga "día" (o "dia"),
"semana", "mes", "trimestre" o "año", así como los plurales.}
}
\description{
Completa una serie de datos
}
