% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NH0358.R
\docType{data}
\name{datos_nh}
\alias{datos_nh}
\alias{NH0358}
\alias{NH0358,NH0114}
\alias{NH0114}
\title{Observaciones diarias de estaciones meteorol칩gicas.}
\format{
NH0358 es un data.table con 25515 filas y 25 columnas

NH0114 es un data.table con 24041 filas y 25 columnas
}
\usage{
NH0358

NH0114
}
\description{
Los datos provienen de 2 estaciones del INTA. Ver \code{\link[=metadatos_nh]{metadatos_nh()}} para m치s
informaci칩n y metadatos de las estaciones.
}
\seealso{
\code{\link[=metadatos_nh]{metadatos_nh()}} devuelve los metadatos de las estaciones meteorol칩gicas.
}
\keyword{datasets}
