% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decil.R
\name{decil}
\alias{decil}
\alias{anomalia_porcentual}
\title{Transformaciones de variables}
\usage{
decil(variable, referencia = rep(TRUE, length(variable)))

anomalia_porcentual(variable, referencia = rep(TRUE, length(variable)))
}
\arguments{
\item{variable}{vector de observaciones de la variable de interés.}

\item{referencia}{serie de precipitación para usar de referencia en el ajuste
a la distribución teórica. Puede ser:
\itemize{
\item vector lógico que se usará para filtrar los datos de entrada.
\item vector numérico de observaciones.
}}
}
\description{
Transformaciones de variables
}
