% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dias-promedios.R
\name{dias_promedio}
\alias{dias_promedio}
\title{Días promedios}
\usage{
dias_promedio(fechas)
}
\arguments{
\item{fechas}{vector de fechas}
}
\description{
Calcula el primer y último día del año promedio para una seria de fechas
}
