% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ith.R
\name{ith}
\alias{ith}
\title{Índice de temperatura y humedad}
\usage{
ith(temperatura, hr)
}
\description{
Calcula el índice de temperatura y humedad (ITH)
}
\details{
@param temperatura vector numérico con valores (o valor) de temperatura en grados
centígrados.
@param hr vector númerico (o valor) de la misma longitud que temperatura con la
humedad relativa en porcentaje.
}
