% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kable.R
\name{kable_inta}
\alias{kable_inta}
\title{Setear tablas en estilo INTA}
\usage{
kable_inta(x, ...)
}
\arguments{
\item{x}{Una tabla.}

\item{...}{Otros argumentos que se pasan a \code{\link[kableExtra:kbl]{kableExtra::kbl()}}}
}
\value{
tabla, objeto kbl.
}
\description{
Llama a \code{\link[kableExtra:kbl]{kableExtra::kbl()}} con defaults apropiados para que siga el estilo
INTA.
}
\examples{

library(dplyr)
library(kableExtra)
metadatos <- metadatos_nh()

metadatos \%>\%
 head() \%>\%
 select(codigo_nh, estacion) \%>\%
 kable_inta(caption = "Ejemplo",
           col.names = c("Código", "Estación")) \%>\%
 kable_styling(latex_options = "scale_down")

}
