% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr.R
\name{knitr_agromet}
\alias{knitr_agromet}
\title{Renderiza un archivo Rmd agregando un resumen}
\usage{
knitr_agromet(archivo, ...)
}
\arguments{
\item{archivo}{La ruta al archivo.}
}
\description{
Renderiza un archivo Rmd agregando un resumen
}
\keyword{internal}
