% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leer_nh.R
\name{leer_nh}
\alias{leer_nh}
\title{Lectura de archivos del sistema NH}
\usage{
leer_nh(archivo)
}
\arguments{
\item{archivo}{caracter o vector de caracteres con el nombre y ubicación de los
archivos a leer}
}
\description{
Lee uno o más archivos
}
