% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leer-surfer.R
\name{leer_surfer}
\alias{leer_surfer}
\title{Lee escalas de Surfer}
\usage{
leer_surfer(archivo, color = c("primario", "secundario"))
}
\arguments{
\item{archivo}{ruta al archivo a laeer.}

\item{color}{caracter que indica qué color leer. Puede ser "primario"
(que corresponde a "FFGColor") o "secundario" (que corresponde a "FBGColor").
No tiene efecto si el archivo es formato LVL1}
}
\value{
Si el archivo es LVL1, un vector con los niveles. Si el archivo es LVL2 o LVL3,
una lista con elementos:
\itemize{
\item \code{niveles} (numérico), el nivel a que corresopnde cada color.
\item \code{colores} (caracter), la representación hexadecimal del color de cada break.
\item \code{paleta} (función), una función que toma un entero \code{n} y devuelve
un vector de caracter con \code{n} colores interpolados a partir de los colores
de la escala.
}
}
\description{
Lee archivos en el formato "Level File Format" de Surfer
(ver http://surferhelp.goldensoftware.com/topics/level_file_format.htm).
}
\examples{
escala <- system.file("extdata", "escala_pp_mensual.lvl", package = "agromet")

escala_pp_mensual <- leer_surfer(escala)

# Valores a los que corresponde cada color
escala_pp_mensual$niveles

# Ver los colores
scales::show_col(escala_pp_mensual$colores)

# Obtener más colores usando la misma paleta
muchos_colores <- escala_pp_mensual$paleta(25)
scales::show_col(muchos_colores)

}
