% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapas.R
\docType{data}
\name{mapa_argentina}
\alias{mapa_argentina}
\alias{mapa_argentina_provincias}
\alias{mapa_argentina_limitrofes}
\title{Mapas}
\format{
A data frame with 53940 rows and 10 variables:
\describe{
\item{price}{price, in US dollars}
\item{carat}{weight of the diamond, in carats}
...
}

An object of class \code{sf} (inherits from \code{data.frame}) with 25 rows and 84 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 7 rows and 64 columns.
}
\source{
rnaturalearth
}
\usage{
mapa_argentina

mapa_argentina_provincias

mapa_argentina_limitrofes
}
\description{
Mapas de Argentina...
}
\keyword{datasets}
