% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapear.R
\name{mapear}
\alias{mapear}
\alias{coord_argentina}
\alias{theme_inta_mapa}
\title{Grafica variables en Argentina}
\usage{
mapear(
  valor,
  lon,
  lat,
  breaks = waiver(),
  escala = scales::viridis_pal(),
  cordillera = FALSE,
  variable = NULL,
  titulo = NULL,
  subtitulo = NULL,
  fuente = NULL
)

coord_argentina(xlim = c(-77, -50), ylim = c(-57, -20), ...)

theme_inta_mapa(...)
}
\arguments{
\item{valor}{vector con los valores medidos.}

\item{lon, lat}{vectores de ubicación en longitud y latitud.}

\item{breaks}{valores donde graficar los contornos. Si es\code{NULL} hace 10 contornos.}

\item{escala}{paleta de colores a usar. Tiene que ser una función que reciba un número
y devuelva esa cantidad de colores. Por ejemplo \link{escala_temp_min}.}

\item{cordillera}{lógico indicando si hay que tapar los datos donde está
la coordillera (donde el kriging es particularmente problemático). Si es \code{TRUE}
pinta con gris donde las alturas son masyores a 1500m. También puede ser un número,
indicando el valor mínimo desde donde empezar a pintar.}

\item{titulo, subtitulo, fuente, variable}{texto para usar como título, subtítulo,
epígrafe y nombre de la guía.}

\item{xlim, ylim}{límites en longitud y latitud.}

\item{...}{otros argumentos que se pasan a \code{\link[ggplot2:ggsf]{ggplot2::coord_sf()}} o \code{\link[ggplot2:ggtheme]{ggplot2::theme_linedraw()}}.}
}
\value{
Un objeto ggplot2.
}
\description{
Dadas mediciones de una veriable en puntos ubicados en Argentina,
interpola a el resto del territorio usando kriging y grafica con contornos
}
\examples{
set.seed(934)
datos_aleatorios <- data.frame(metadatos_nh(),
                               pp = rgamma(nrow(metadatos_nh()), 0.5, scale = 1)*25)

with(datos_aleatorios, mapear(pp, lon, lat, cordillera = TRUE,
                              escala = escala_pp_diaria,
                              variable = "mm",
                              titulo = "Precipitación aleatoria",
                              fuente = "Fuente: datos de ejemplo"))

}
