% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leer-nh.R
\name{metadatos_nh}
\alias{metadatos_nh}
\title{Tabla de Metadatos de Estaciones NH}
\usage{
metadatos_nh(
  codigo = NULL,
  provincia = NULL,
  organismo = NULL,
  lat = NULL,
  lon = NULL
)
}
\arguments{
\item{codigo, provincia, organismo}{carácter o vector de caracteres para filtrar según código de estación, provincia u organismo.}

\item{lat}{vector numérico con las latitudes límite de la región de interés.}

\item{lon}{vector numérico con las longitudes límite de la región de interés
(entre -180 y 180).}
}
\value{
data.frame con los metadatos de las estaciones cuyas columnas incluye:
\itemize{
\item \code{codigo_nh} (caracter), variable llave para acceder a los metadatos de estaciones
\item \code{estacion} (caracter), nombre de la estación
\item \code{provincia} (caracter), provincia donde se encuentra la estación
\item \code{organismo} (caracter), organismo a cargo de la estación
\item \code{lat} (numérico), longitud
\item \code{lon} (numérico), latitud
\item \code{altura} (numérico), altura sobre el nivel del mar de la estación
}
}
\description{
Devuelve los metadatos de estaciones incluyendo el código único, ubicación (latitud y
longitud) y el nombre.
}
\details{
Esta función por defecto devuelve la lista completa de estaciones pero alternativamente
se puede devolver estaciones específicas a partir de sus códigos o todas las estaciones
incluidas en una región. Además incluye un método plot para visualizar rápidamente
la ubicación de las estaciones.
}
\examples{
# listado completo de estaciones
head(metadatos_nh(), n = 10)

# listado de estaciones específicas
metadatos_nh(codigo = c("0001", "0011"))

# Filtrar por provincias
metadatos_nh(provincia = c("La Pampa", "Catamarca"))

# Filtrar por organismo
metadatos_nh(organismo = "INTA")

# listados de estaciones en una región
metadatos_nh(lat = c(-30, -20), lon = c(-65, -55))

# gráfico
plot(metadatos_nh())

}
