% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NH0011.R
\docType{data}
\name{NH0011}
\alias{NH0011}
\title{Datos diarios de estación}
\format{
An object of class \code{data.frame} with 23524 rows and 25 columns.
}
\usage{
NH0011
}
\description{
Datos diarios de estación
}
\keyword{datasets}
