% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NH0358.R
\docType{data}
\name{NH0358}
\alias{NH0358}
\title{Datos diarios de estación}
\format{
An object of class \code{data.frame} with 25515 rows and 25 columns.
}
\usage{
NH0358
}
\description{
Datos diarios de estación
}
\keyword{datasets}
