% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olas.R
\name{olas}
\alias{olas}
\title{Olas}
\usage{
olas(fecha, ..., remplaza.na = FALSE)
}
\arguments{
\item{fecha}{vector de fechas, la serie temporal debe estar completa, sin datos
faltantes implicitos.}

\item{...}{umbral o umbrales a calcular utilizando operadores lógicos.}

\item{remplaza.na}{lógico. Por defecto es FALSE, es decír que si la función
encuentra un dato faltante "corta" la ola o periodo de persitencia. Si es
TRUE, la función reemplaza cada NA por el valor previo en la serie, por lo tanto
la ola no se interrumpe si hay NAs.}
}
\value{
Devuelve un data.frame con 3 variables fijas y las posibles variables
asociadas al agrupamiento:
\itemize{
\item \code{ola} (caracter) nombre de la ola definido por el usuario
(si los argumentos de \code{...} no tienen nombre, se usa \code{V1}, \code{V2}, etc...)
\item \code{inicio} (fecha) fecha de inicio de la ola o periodo de persistencia
\item \code{fin} (fecha) fecha de finalización de la ola o periodo de persistencia
\item \code{duracion} (diferencia de fechas, tipo drtn) duración de la ola
}

Si una ola todavía no terminó, fin y longitud son NA.
}
\description{
Identifica periodos de persistencia de un evento definido a partir de alguna
condición lógica, por ejemplo días consecutivos donde la temperatura mínima
fue igual o menor a 0°C para calcular días acumulados de heladas.
}
\details{
La función Puede utilizarse en el contexto de \code{\link[dplyr:summarise]{dplyr::summarise()}} y \code{\link[dplyr:group_by]{dplyr::group_by()}}
para hacer este cálculo por grupos.
}
\examples{
data(NH0358)

library(dplyr)
NH0358 \%>\%
  reframe(olas(fecha, calor = t_max > 20)) \%>\%
  slice_head(n = 10)

NH0358 \%>\%
  reframe(olas(fecha, frio = t_min <= 0)) \%>\%
  slice_head(n = 10)

}
