% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decil.R
\name{porcentaje_normal}
\alias{porcentaje_normal}
\title{Tranforma precipitación en porcentaje con respecto a la precipitación normal}
\usage{
porcentaje_normal(
  precipitacion,
  pp_referencia = precipitacion,
  normalizado = FALSE
)
}
\arguments{
\item{precipitacion}{Vector de observaciones de precipitacion}

\item{pp_referencia}{Vector de observacione sprecipitación para usar de
referencia al calcular el valor normal. Por defecto usa todos los
datos.}

\item{normalizado}{Valor lógico para indicar si los deciles se devuelven
normalizados según (porcentaje - 1).}
}
\description{
Tranforma precipitación en porcentaje con respecto a la precipitación normal
}
