% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapear.R
\name{scale_fill_precipitacion_d}
\alias{scale_fill_precipitacion_d}
\alias{breaks_precipitacion}
\alias{paleta_precipitacion}
\title{Escalas para precipitacion}
\usage{
scale_fill_precipitacion_d(
  name = waiver(),
  breaks = waiver(),
  drop = !is.vector(breaks),
  palette = paleta_precipitacion,
  ...
)

breaks_precipitacion()

paleta_precipitacion(n)
}
\arguments{
\item{name}{Nombre de la escala.}

\item{breaks}{Cortes o valores de la escala.}

\item{drop}{Lógico que indica si se muestran todos los valores o sólo los
presentes en los datos.}

\item{palette}{Paleta a usar. Una función que recibe un parámetro (el número de colores
deseados) y devuelve un vector con colores.}

\item{n}{El número de colores deseados}
}
\description{
Escalas para precipitacion
}
