% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-inta.R
\name{scale_fill_inta}
\alias{scale_fill_inta}
\alias{scale_color_inta}
\title{Escalas de colores para precipitación y temperatura}
\usage{
scale_fill_inta(
  escala,
  name = waiver(),
  breaks = waiver(),
  drop = waiver(),
  ...
)

scale_color_inta(
  escala,
  name = waiver(),
  breaks = waiver(),
  drop = waiver(),
  ...
)
}
\arguments{
\item{escala}{escala de colores. Puede ser
\itemize{
\item una lista con elementos \code{niveles} y \code{paleta} (ver \verb{[leer_surfer()]} y \link{escala_temp_min}).
\item una función que toma un entero \code{n} y devuelve
un vector de caracter con \code{n} colores interpolados a partir de los colores
de la escala.)
}}

\item{name}{nombre de la escala.}

\item{breaks}{niveles de la escala. Si no es \code{waiver()}, tiene prioridad por sobre
los niveles definidos en \code{escala}.}

\item{drop}{lógico que indica si se muestran todos los valores o sólo los
presentes en los datos. Por defecto, es \code{FALSE} si la escala define los niveles usando
\code{breaks} o \code{escala}.}

\item{...}{otros argumentos que se pasan a \code{\link[ggplot2:scale_manual]{ggplot2::scale_fill_manual()}} o \code{\link[ggplot2:discrete_scale]{ggplot2::discrete_scale()}}.}
}
\value{
objeto ggproto compatible con ggplot2.
}
\description{
Escalas para \code{color} y \code{fill} para variables discretas.
}
\examples{
library(ggplot2)
library(dplyr)

pp_enero <- datos_nh_mensual |>
  filter(mes == unique(mes)[1])

# Los contornos llenos requieren que los datos estén en una grilla
# regular, necesitamos hacer una interpolación con kriging.
with(pp_enero, agroclimatico:::kringe(precipitacion_mensual, lon, lat)) |>
ggplot(aes(lon, lat)) +
 geom_contour(aes(z = var1.pred)) +
 geom_contour_filled(aes(z = var1.pred)) +
 scale_fill_inta(escala = escala_pp_mensual)

}
