% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spi.R
\name{spi}
\alias{spi}
\alias{spei}
\alias{spi_referencia}
\title{Calcula el SPI}
\usage{
spi(
  fecha,
  precipitacion,
  escalas,
  referencia = spi_referencia(fecha, precipitacion),
  distribucion = "Gamma",
  ...
)

spei(fecha, balance, escalas, distribucion = "log-Logistic", ...)

spi_referencia(fecha, precipitacion)
}
\arguments{
\item{fecha}{Vector de fechas.}

\item{precipitacion}{Vector de precipitacion.}

\item{escalas}{Vector numérico con las escalas requeridas}

\item{distribucion}{Distribución usada para ajustar los datos.}

\item{...}{Argumentos pasados a \link[SPEI:Drought-indices]{SPEI::spi}}

\item{balance}{Balance entre precipitación y evapotaranspiración potencial.}
}
\description{
Calcula el SPI
}
