% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agromet-package.R
\docType{package}
\name{agromet-package}
\alias{agromet}
\alias{agromet-package}
\title{agromet: Lo Que Hace El Paquete}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Más sobre el paquete.
}
\author{
\strong{Maintainer}: Elio Campitelli \email{elio.campitelli@cima.fcen.uba.ar} (\href{https://orcid.org/0000-0002-7742-9230}{ORCID})

Authors:
\itemize{
  \item Paola Corrales \email{paola.corrales@cima.fcen.uba.ar}
}

}
\keyword{internal}
