% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decil.R
\name{anomalia_porcentual}
\alias{anomalia_porcentual}
\title{Tranforma una variable en porcentaje con respecto a sus valores normales.}
\usage{
anomalia_porcentual(
  precipitacion,
  pp_referencia = precipitacion,
  normalizado = TRUE
)
}
\arguments{
\item{normalizado}{Valor lógico para indicar si los deciles se devuelven
normalizados según (porcentaje - 1).}

\item{variable}{Vector de observaciones de la variable de interés.}

\item{variable_referencia}{Vector de observacione de la variable para usar de
referencia al calcular el valor normal. Por defecto usa todos los
datos.}
}
\description{
Tranforma una variable en porcentaje con respecto a sus valores normales.
}
