% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/completar-serie.R
\name{completar_serie}
\alias{completar_serie}
\title{Completa una Serie de Datos}
\usage{
completar_serie(datos, fecha, resolucion, rango = range(fecha))
}
\arguments{
\item{datos}{tabla (data.frame, data.table, tibble) a completar.}

\item{fecha}{variable tipo fecha (Date, IDate, POSIXct, etc).}

\item{resolucion}{texto que define resolución de salida de los datos. Puede ser
cualquier valor aceptado por el argumento \code{by} de la función \code{\link[=seq.Date]{seq.Date()}}
o su traducción al español. Es decir, "día" (o "dia"),
"semana", "mes", "trimestre" o "año", así como los plurales.}

\item{rango}{un vector cuyo rango define el período a completar. Es útil si
se quiere que múltiples grupos de datos tengan el mismo rango de fechas. Es
posible definir un rango por fuera del rango original de los datos para homogeneizar
series temporales.}
}
\value{
Devuelve un data.frame con las mismas variables de origen. La variable
asociada a las fechas ahora se encuentra completa para la resolución indicada
y el resto de las variables se completan con NA.
}
\description{
La función permite completar una serie de datos temporales definiendo alguna
resolución disponible. Es compatible con datos agrupados por \code{\link[dplyr:group_by]{dplyr::group_by()}}.
}
\examples{
# Datos de prueba completos
datos <- data.frame(fechas = seq(as.Date("1985-01-01"), as.Date("2015-12-01"),
                             by = "1 month"),
                    pp = 1)

set.seed(42)
datos_perdidos <- sample(nrow(datos), nrow(datos)/10)
datos$pp[datos_perdidos] <- NA
datos_incompletos <- na.omit(datos) #Serie de datos a completar

completar_serie(datos_incompletos, fechas, resolucion = "1 mes")

}
