% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decil.R
\name{decil}
\alias{decil}
\alias{anomalia_porcentual}
\title{Transformaciones de variables}
\usage{
decil(variable, variable_referencia = variable)

anomalia_porcentual(variable, variable_referencia = variable)
}
\arguments{
\item{variable}{Vector de observaciones de la variable de interés.}

\item{variable_referencia}{Vector de observacione de la variable de interés para usar de
referencia al calcular la transformación. Por defecto usa todos los
datos incluidos en el argumento variable.}
}
\description{
Transformaciones de variables
}
