% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decil.R
\name{decil}
\alias{decil}
\title{Transforma precipitación en deciles}
\usage{
decil(precipitacion, pp_referencia = precipitacion, normalizado = FALSE)
}
\arguments{
\item{precipitacion}{Vector de observaciones de precipitacion}

\item{pp_referencia}{Vector de observacione sprecipitación para usar de
referencia al calcular el valor de los deciles. Por defecto usa todos los
datos.}

\item{normalizado}{Valor lógico para indicar si los deciles se devuelven
normalizados según (deciles - 5)/10 para que devuelva valores entre -0.5 y 0.5
en vez de entre 0 y 10.}
}
\description{
Transforma precipitación en deciles
}
