% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decil.R
\name{decil}
\alias{decil}
\alias{anomalia_porcentual}
\title{Transformaciones estadísticas de variables}
\usage{
decil(variable, referencia = rep(TRUE, length(variable)))

anomalia_porcentual(
  variable,
  referencia = rep(TRUE, length(variable)),
  na.rm = FALSE
)
}
\arguments{
\item{variable}{vector de observaciones de la variable de interés.}

\item{referencia}{serie de observaciones para usar de referencia en el ajuste
a la distribución teórica. Puede ser:
\itemize{
\item vector lógico que se usará para filtrar los datos de entrada.
\item vector numérico de observaciones.
\item la serie completa (opción por defecto).
}}

\item{na.rm}{lógico. Define si se utilizan o no valores faltantes en el cálculo
de la anomalía porcentual.}
}
\value{
Para el cálculo de los deciles, devuelve un vector numérico con el decil
asociado a cada valor de la variable. En este caso la columna \code{deciles} es de
tipo doble ya que devuelve el valor exacto del decil sin redondeos.
Para el cálculo de la anomalía porcentual también
devuelve un vector numérico. Las funciones son compatibles con
\code{\link[dplyr:group_by]{dplyr::group_by()}} y \code{\link[dplyr:mutate]{dplyr::mutate()}}.
}
\description{
Las funciones \code{decil()} y \code{anomalia_porcentual()} devuelven estos estadísticos
para alguna variable dado un periodo de referencia especificado.
}
\details{
Para recuperar el valor de la variable asociado a determinado decil se puede
utilizar la función \code{\link[stats:quantile]{stats::quantile()}}.
}
\examples{
library(dplyr)
data(NH0358)

# Deciles de precipitación usando como referencia la serie completa
NH0358 \%>\%
  mutate(deciles = decil(precip)) \%>\%
  slice_head(n = 10)

# Deciles mensuales
precip_mensual <- NH0358 \%>\%
  group_by(fecha = lubridate::floor_date(fecha, "month")) \%>\%
  summarise(precip = sum(precip, na.rm = TRUE))

precip_mensual \%>\%
  mutate(deciles = decil(precip))

# Deciles definiendo un periodo de referencia
precip_mensual \%>\%
  mutate(deciles = decil(precip,
                         referencia = lubridate::year(fecha) <= 1958))

# Anomalia porcentual usando como referencia la serie completa
NH0358 \%>\%
  mutate(anomalia = anomalia_porcentual(precip)) \%>\%
  slice_head(n = 10)

#  Anomalia porcentual definiendo un periodo de referencia
precip_mensual \%>\%
  mutate(deciles = anomalia_porcentual(precip,
                         referencia = lubridate::year(fecha) <= 1958))

}
