% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decil.R
\name{decil}
\alias{decil}
\alias{anomalia_porcentual}
\title{Transformaciones estadísticas de variables}
\usage{
decil(variable, referencia = rep(TRUE, length(variable)))

anomalia_porcentual(
  variable,
  referencia = rep(TRUE, length(variable)),
  na.rm = FALSE
)
}
\arguments{
\item{variable}{vector de observaciones de la variable de interés.}

\item{referencia}{serie de observaciones para usar de referencia en el ajuste
a la distribución teórica. Puede ser:
\itemize{
\item vector lógico que se usará para filtrar los datos de entrada.
\item vector numérico de observaciones.
\item la serie completa (opción por defecto).
}}

\item{na.rm}{lógico. Define si se utilizan o no valores faltantes en el cálculo
de la anomalía porcentual.}
}
\value{
Devuelve un data.frame con el valor del decil o la anomalía porcentual
según sea el caso. Es compatible con \code{\link[dplyr:group_by]{dplyr::group_by()}} y \code{\link[dplyr:mutate]{dplyr::mutate()}}.
}
\description{
Las funciones \code{decil()} y \code{anomalia_porcentual()} devuelven estos estadísticos
para alguna variable dado un periodo de referencia especificado.
}
\examples{
library(dplyr)
data(NH0358)

# Deciles de precipitación usando como referencia la serie completa
NH0358 \%>\%
  mutate(deciles = decil(precip))

# Deciles mensuales
precip_mensual <- NH0358 \%>\%
  group_by(fecha = lubridate::floor_date(fecha)) \%>\%
  summarise(precip = mean(precip, na.rm = TRUE))

precip_mensual \%>\%
  mutate(deciles = decil(precip))

# Definiendo un periodo de referencia
precip_mensual \%>\%
  mutate(deciles = decil(precip,
                         referencia = lubridate::year(fecha) <= 1958))

}
