% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dias-promedio.R
\name{dias_promedio}
\alias{dias_promedio}
\title{Días Promedio}
\usage{
dias_promedio(fechas)
}
\arguments{
\item{fechas}{vector de fechas}
}
\value{
La función devuelve un data.frame con 4 variables fijas y variables extras
en el caso de hacer el cálculo para distintos grupos:
\itemize{
\item \code{variable} (caracter) primer_dia o ultimo_dia según corresponda
\item \code{dia} (numérico) día del mes
\item \code{mes} (numérico) mes de ocurrencia
\item \code{dia_juliano} (numérico) día del año
}
}
\description{
Calcula el primer y último día del año promedio a partir de una seria de fechas.
}
\details{
Esta función solo requiere un vector de fechas para calcular el primer y último
día del año en promedio. Si este vector incluye todos los días de muchos años
el resultado será el 1° de enero y el 31 de diciembre. Pero si solo se utilizan
las fechas que cumplen con una determinada condición, por ejemplo aquellos
días donde la temperatura mínima fue menor o igual a 0°C, entonces devuelve
el primer y último día de ocurrencia en promedio para este evento.
}
\examples{
data(NH0358)

# Usando la serie completa
dias_promedio(NH0358$fecha)

# Filtrando los datos para un determinado evento
library(dplyr)
NH0358 \%>\%
  filter(t_min <= 0) \%>\%
  summarise(dias_promedio(fecha))

}
