% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/escalas.R
\docType{data}
\name{escala_temp_min}
\alias{escala_temp_min}
\alias{escala_temp_max}
\alias{escala_pp_mensual}
\alias{escala_pp_diaria}
\title{Escalas de colores usadas por el INTA}
\format{
Listas en el mismo formato que devuelve \code{leer_surfer()} con elementos:
\itemize{
\item \code{niveles} (numérico), el nivel a que corresopnde cada color.
\item \code{colores} (caracter), la representación hexadecimal del color de cada break.
\item \code{paleta} (función), una función que toma un entero \code{n} y devuelve
un vector de caracter con \code{n} colores interpolados a partir de los colores
de la escala.
}
}
\usage{
escala_temp_min

escala_temp_max

escala_pp_mensual

escala_pp_diaria
}
\description{
Escalas de colores típicas usadas por INTA para distintas variables.
}
\keyword{datasets}
