% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ith.R
\name{ith}
\alias{ith}
\title{Índice de Temperatura y Humedad}
\usage{
ith(temperatura, hr)
}
\arguments{
\item{temperatura}{vector numérico con valores (o valor) de temperatura en grados
centígrados.}

\item{hr}{vector númerico (o valor) de la misma longitud que temperatura con la
humedad relativa en porcentaje.}
}
\value{
Devuelve un valor o vector de valores con el ITH. Este valor es utilizado
como una medida de la intensidad de las condiciones de estrés por calor a la
que se encuentra expuesto el animal. Para bovinos se categoriza como:
\itemize{
\item \strong{Normal} si ITH < 75
\item \strong{Alerta} para ITH entre 75 y 78
\item \strong{Peligro} para ITH entre 79 y 83
\item \strong{Emergencia} para ITH >= 84
}
}
\description{
Calcula el índice de temperatura y humedad (ITH)
}
\examples{
ith(temperatura = 23, hr = 65)

data(NH0358)

# En el contexto de mutate
library(dplyr)
NH0358 \%>\%
  filter(!is.na(hr)) |>
  mutate(t_media = (t_max + t_min)/2) \%>\%
  mutate(ith = ith(t_media, hr)) \%>\%
  slice_head(n = 10)

}
\references{
\itemize{
\item Armendano, J. I. ¿Cuándo se generan condiciones de estrés por calor en bovinos para carne?
\href{https://inta.gob.ar/sites/default/files/inta_estres_por_calor_bovinos_para_carne.pdf}{link}
\item Armstrong, DV. 1994. Heat stress interaction with shade and cooling. J. Diary Sci. 77:2004-2050
}
}
