% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leer-nh.R
\name{leer_nh}
\alias{leer_nh}
\title{Lectura de Archivos con Formato NH}
\usage{
leer_nh(archivos)
}
\arguments{
\item{archivos}{Caracter o vector de caracteres con nombre y ubicación de los
archivos a leer.}
}
\value{
Devuelve un data.frame con tantas filas como líneas en el o los archivos leidos y
las 25 variables presentes.
}
\description{
Lee uno o más archivos siempre que mantengan el formato NH de ancho fijo.
}
\details{
La función está preparada para leer datos díarios con el formato NH que incluye
25 variables:
\itemize{
\item \code{codigo} (caracter)
\item \code{codigo_nh} (caracter), variable llave para acceder a los metadatos de estaciones
\item \code{fecha} (fecha)
\item \code{t_max} (numérico), temperatura máxima en grados centígrados
\item \code{t_min} (numérico), temperatura mínima en grados centígrados
\item \code{precip} (numérico), precipitación acumulada en milímetros
\item \code{lluvia_datos} (numérico), ocurrencia de precipitación 1 indica lluvia
\item \code{lluvia} (numérico),  ocurrencia de lluvia
\item \code{llovizna} (numérico), ocurrencia de llovizna
\item \code{granizo} (numérico), ocurrencia de granizo
\item \code{nieve} (numérico), ocurrencia de nieve
\item \code{t_min_5cm} (numérico), temperatura mínima a intemperie a 5cm en grados centígrados
\item \code{t_min_50cm} (numérico), temperatura mínima a intemperie a 50cm en grados centígrados
\item \code{t_suelo_5cm} (numérico), temperatura media del suelo a 5cm en grados centígrados
\item \code{t_suelo_10cm} (numérico), temperatura media del suelo a 10cm en grados centígrados
\item \code{heliofania_efec} (numérico), heliofanía efectiva en horas
\item \code{heliofania_rel} (numérico), heliofanía relativa en porcentaje
\item \code{p_vapor} (numérico), tensión de vapor en hPa
\item \code{hr} (numérico), humedad relativa en porcentaje
\item \code{td} (numérico), temperatura de rocío en grados centígrados
\item \code{rocio} (numérico), ocurrencia de rocío
\item \code{viento_10m} (numérico), viento a 10 metros en km/h
\item \code{viento_2m} (numérico), viento a 2 metros en km/h
\item \code{rad} (numérico), radiación en MJ/m2
\item \code{etp} (numérico), evapotranspiración en milímetros
}
}
\examples{
archivo <- system.file("extdata", "NH0358.DAT", package = "agroclimr")
datos <- leer_nh(archivo)

}
\seealso{
\code{\link[=metadatos_nh]{metadatos_nh()}} devuelve los metadatos de las estaciones meteorológicas.
}
