% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapas.R
\name{mapa_argentina}
\alias{mapa_argentina}
\alias{mapa_provincias}
\alias{mapa_argentina_limitrofes}
\alias{mapa_departamentos}
\title{Mapas}
\usage{
mapa_argentina()

mapa_provincias(provincias = NULL, departamentos = FALSE)

mapa_argentina_limitrofes()

mapa_departamentos(provincias = NULL)
}
\arguments{
\item{provincias}{vector de caracteres con los nombres de provincias a filtrar.
Si es \code{NULL}, devuelve todas las provincias de Argentina.}

\item{departamentos}{lógico. Si es \code{TRUE} grafica los departamentos.}
}
\value{
Devuelve una tibble con las variables necesarias para generar un mapa
utilizando \link{ggplot2} y \link{sf}.
}
\description{
Mapas de Argentina, sus provincias, departamentos y paises limítrofes. Los
mapas de departamentos surgen del repositorio
público del \href{https://www.ign.gob.ar/NuestrasActividades/InformacionGeoespacial/CapasSIG}{Instituto Geográfico Nacional},
mientas que el mapa de países limítrofes, Argentina y sus provincias son parte del repositorio
\href{https://www.naturalearthdata.com/}{Natural Earth}.
}
\examples{
library(ggplot2)

# Solo Argentina
ggplot() +
  geom_sf(data = mapa_argentina())

# Argentina y sus provincias
ggplot() +
  geom_sf(data = mapa_provincias())

# Algunas provincias
ggplot() +
  geom_sf(data = mapa_provincias(provincias = c("La Pampa", "Córdoba")))

# Algunas provincias y sus departamentos
ggplot() +
  geom_sf(data = mapa_provincias(provincias = c("La Pampa", "Córdoba"),
                                 departamentos = TRUE))

}
