% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapas.R
\name{mapa_argentina}
\alias{mapa_argentina}
\alias{mapa_provincias}
\alias{mapa_argentina_limitrofes}
\alias{mapa_departamentos}
\title{Mapas}
\usage{
mapa_argentina()

mapa_provincias(provincias = NULL, departamentos = FALSE)

mapa_argentina_limitrofes()

mapa_departamentos(provincias = NULL)
}
\arguments{
\item{provincias}{Vector de provincias para filtrar. Si es \code{NULL},
no filtra.}

\item{departamentos}{Lógico indicando si dibujar los departamenteos o no.}
}
\description{
Mapas
}
