% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapear.R
\name{mapear}
\alias{mapear}
\title{Grafica variables en Argentina}
\usage{
mapear(
  valor,
  lon,
  lat,
  breaks = NULL,
  paleta = scales::viridis_pal(),
  cordillera = FALSE,
  titulo = NULL,
  subtitulo = NULL,
  fuente = NULL
)
}
\arguments{
\item{valor}{vector con los valores medidos.}

\item{lon, lat}{vectores de ubicación en longitud y latitud.}

\item{breaks}{valores donde graficar los contornos. Si es\code{NULL} hace 10 contornos.}

\item{paleta}{paleta de colores a usar. Tiene que ser una función que reciba un número
y devuelva esa cantidad de colores. Por ejemplo \code{\link[=paleta_precipitacion]{paleta_precipitacion()}}.}

\item{cordillera}{lógico indicando si hay que tapar los datos donde está
la coordillera (donde el kriging es particularmente problemático). Si es \code{TRUE}
pinta con gris donde las alturas son masyores a 1500m. También puede ser un número,
indicando el valor mínimo desde donde empezar a pintar.}

\item{titulo, subtitulo, fuente}{texto para usar como título, subtítulo y
epígrafe.}
}
\value{
Un objeto ggplot2.
}
\description{
Dadas mediciones de una veriable en puntos ubicados en Argentina,
interpola a el resto del territorio usando kriging y grafica con contornos
}
\examples{
set.seed(934)
datos_aleatorios <- data.frame(metadatos_nh(), pp = rgamma(nrow(metadatos_nh()), 1, scale = 5))

with(datos_aleatorios, mapear(pp, lon, lat, cordillera = TRUE,
                              paleta = paleta_precipitacion,
                              titulo = "Precipitación aleatoria",
                              fuente = "Fuente: datos de ejemplo"))

}
