% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapear.R
\name{mapear}
\alias{mapear}
\alias{coord_argentina}
\alias{theme_inta_mapa}
\title{Grafica variables en Argentina}
\usage{
mapear(
  data,
  valor,
  lon,
  lat,
  breaks = waiver(),
  escala = scales::viridis_pal(),
  cordillera = FALSE,
  variable = NULL,
  titulo = NULL,
  subtitulo = NULL,
  fuente = NULL
)

coord_argentina(xlim = c(-77, -50), ylim = c(-57, -20), ...)

theme_inta_mapa(...)
}
\arguments{
\item{valor}{vector con los valores medidos.}

\item{lon, lat}{vectores de ubicación en longitud y latitud.}

\item{breaks}{vector numérico que define para que valores se graficará los
contornos. Si es\code{NULL} hace 10 contornos calculados a partir el rango de los datos.}

\item{escala}{paleta de colores a usar. Tiene que ser una función que reciba un número
y devuelva esa cantidad de colores. Por ejemplo \link{escala_temp_min}.}

\item{cordillera}{valor lógico indicando si hay que tapar los datos donde está
la coordillera (donde el kriging es particularmente problemático). Si es \code{TRUE}
pinta con gris donde la altura de la topografía es mayor a 1500 m. También
puede ser un número, indicando el valor mínimo desde donde empezar a graficar
la cordillera.}

\item{titulo, subtitulo, fuente, variable}{texto para usar como título, subtítulo,
epígrafe y nombre de la leyenda.}

\item{xlim, ylim}{límites en longitud y latitud.}

\item{...}{otros argumentos que se pasan a \code{\link[ggplot2:ggsf]{ggplot2::coord_sf()}} o \code{\link[ggplot2:ggtheme]{ggplot2::theme_linedraw()}}.}
}
\value{
Un objeto ggplot2.
}
\description{
Dadas mediciones de una variable en puntos ubicados en Argentina,
interpola al resto del territorio usando kriging y grafica con contornos
Las funciones secundarias \code{\link[=coord_argentina]{coord_argentina()}} y \code{\link[=theme_inta_mapa]{theme_inta_mapa()}} permiten
generar un mapa en la región de Argentina (definida por \code{xlim} y \code{ylim}) con
el estilo específico usando por INTA.
}
\examples{
\dontrun{

data(datos_nh_mensual)

with(datos_nh_mensual |> filter(mes == unique(mes)[4]), kringe(precipitacion_mensual, lon, lat)) |>
ggplot(aes(lon, lat)) +
  geom_contour(aes(z = var1.pred)) +
  geom_contour_filled(aes(z = var1.pred)) +
  scale_fill_inta(escala = escala_pp_mensual)
}

}
