% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olas.R
\name{olas}
\alias{olas}
\title{Olas}
\usage{
olas(fecha, condicion)
}
\arguments{
\item{fecha}{vector de fechas.}

\item{condicion}{vector lógico.}
}
\value{
Devuelve un data.frame con 3 variables fijas y las posibles variables
asociadas al agrupamiento:
\itemize{
\item \code{inicio} (fecha) fecha de inicio de la ola o periodo de persistencia
\item \code{fin} (fecha) fecha de finalización de la ola o periodo de persistencia
\item \code{longitud} (diferencia de fechas, drtn) duración de la ola
}

Si una ola todavía no terminó, su día de fin es NA y su longitud es NA.
}
\description{
Identifica periodos de persistencia de un evento definido a partir de alguna
condición lógica, por ejemplo días consecutivos donde la temperatura mínima
fue igual o menor a 0°C para calcular días acumulados de heladas.
}
\details{
La función es sensible a los datos faltantes, esto quiere decir que si se
encuentra con un dato faltante la función "corta" el periodo de persistencia.
Puede utilizarse en el contexto de \code{\link[=summarise]{summarise()}} y \code{\link[=group_by]{group_by()}} para hacer
este cálculo por grupos.
}
\examples{
archivo <- system.file("extdata", "NH0011.DAT", package = "agromet")
datos <- leer_nh(archivo)

library(dplyr)
datos \%>\%
  summarise(olas(fecha, t_max > 20))

}
