% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-inta.R
\name{scale_fill_inta}
\alias{scale_fill_inta}
\alias{scale_color_inta}
\title{Escalas para precipitacion}
\usage{
scale_fill_inta(
  escala,
  name = waiver(),
  breaks = waiver(),
  drop = waiver(),
  ...
)

scale_color_inta(
  escala,
  name = waiver(),
  breaks = waiver(),
  drop = waiver(),
  ...
)
}
\arguments{
\item{escala}{escala de colores. Puede ser
\itemize{
\item una lista con elementos \code{niveles} y \code{paleta} (ver \verb{[leer_surfer()]} y \link{escala_temp_min}).
\item una función que toma un entero \code{n} y devuelve
un vector de caracter con \code{n} colores interpolados a partir de los colores
de la escala.)
}}

\item{name}{nombre de la escala.}

\item{breaks}{niveles de la escala. Si no es \code{waiver()}, tiene prioridad por sobre
los niveles deinidos en \code{escala}.}

\item{drop}{lógico que indica si se muestran todos los valores o sólo los
presentes en los datos. Por defecto, es \code{FALSE} si la escala define los niveles usando
\code{breaks} o \code{escala}.}

\item{...}{otros argumentos que se padan a \code{\link[ggplot2:scale_manual]{ggplot2::scale_fill_manual()}} o \code{\link[ggplot2:discrete_scale]{ggplot2::discrete_scale()}}.}
}
\value{
objeto ggproto compatible con ggplot2.
}
\description{
Escalas para precipitacion
}
