% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spi.R
\name{spi}
\alias{spi}
\title{Calcula el SPI}
\usage{
spi(fecha, precipitacion, escalas, ...)
}
\arguments{
\item{fecha}{Vector de fechas.}

\item{precipitacion}{Vector de precipitacion.}

\item{escalas}{Vector numérico con las escalas requeridas}

\item{...}{Argumentos pasados a \link[SPEI:spei]{SPEI::spi}}
}
\description{
Calcula el SPI
}
