% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spi.R
\name{spi}
\alias{spi}
\title{Calcula el SPI}
\usage{
spi(fecha, precipitacion, k)
}
\arguments{
\item{fecha}{Vector de fechas.}

\item{precipitacion}{Vector de precipitacion.}

\item{k}{Vector numérico con las escalas requeridas}
}
\description{
Calcula el SPI
}
